+++ src/assets/index.ts
@@ -1,13 +1,19 @@
 import creditsafe from "./company_logos/creditsafe.png";
 import financialforce from "./company_logos/financialforce.png";
-import { prayerpal } from "./project_screenshots/prayerpal/index";
-import { instock } from "./project_screenshots/instock/index";
+import { instock } from "./project_screenshots/instock";
 import { wrackreport } from "./project_screenshots/wrackreport";
+import { prayerpal } from "./project_screenshots/prayerpal";
+import { e2s } from "./project_screenshots/e2s";
+import { hafod } from "./project_screenshots/hafod";
+import wave from "./loading/wave.png";
 
 export {
-    creditsafe,
-    financialforce,
-    prayerpal,
-    instock,
-    wrackreport
-}
+  creditsafe,
+  financialforce,
+  instock,
+  wrackreport,
+  prayerpal,
+  e2s,
+  hafod,
+  wave,
+};