+++ src/assets/project_screenshots/e2s/index.ts
@@ -1,0 +1,25 @@
+import dashboard from "./dashboard.png";
+import reports from "./reports.png";
+import fullReport from "./full_report.png";
+import costForecast from "./cost_forecast.png";
+import billValidation from "./bill_validation.png";
+import overview from "./overview.png";
+import userManagement from "./user_management.png";
+import addUser from "./add_user.png";
+import siteManagement from "./site_management.png";
+import addSite from "./add_site.png";
+import deleteSite from "./delete_site.png";
+
+export const e2s = [
+  dashboard,
+  reports,
+  fullReport,
+  costForecast,
+  billValidation,
+  overview,
+  userManagement,
+  addUser,
+  siteManagement,
+  addSite,
+  deleteSite,
+];