+++ src/assets/project_screenshots/hafod/index.ts
@@ -1,0 +1,17 @@
+import login from "./login.png";
+import home from "./home.png";
+import charts from "./charts.png";
+import footer from "./footer.png";
+import tenants from "./tenants.png";
+import tenantSearch from "./tenant_search.png";
+import addTenant from "./add_tenant.png";
+
+export const hafod = [
+    login,
+    home,
+    charts,
+    footer,
+    tenants,
+    tenantSearch,
+    addTenant,
+]