+++ src/pages/index.ts
@@ -1,7 +1,9 @@
 import Navigation from "./navigation/Navigation";
+import Home from "./Home";
 import Landing from "./landing/Landing";
 import About from "./about/About";
 import Experiences from "./experience/Experiences";
 import Projects from "./projects/Projects";
+import ProjectsList from "./projects/ProjectsList";
 
-export { Navigation, Landing, About, Experiences, Projects };
+export { Navigation, Home, Landing, About, Experiences, Projects, ProjectsList };