+++ src/pages/projects/Projects.tsx
@@ -3,6 +3,10 @@ import { slideDown } from "../../utils/motionVariants";
 import SectionWrapper from "../../utils/SectionWrapper";
 import { projects } from "../../utils/constants";
 import ProjectCard from "./ProjectCard";
+import IconButton from "../../components/IconButton";
+import { useNavigate } from "react-router-dom";
+import { IconArrowRight } from "@tabler/icons-react";
+import { Project } from "../../utils/types";
 
 type ProjectsProp = {
   isDarkMode: boolean;
@@ -10,6 +14,8 @@ type ProjectsProp = {
 };
 
 function Projects({ isMobile, isDarkMode }: ProjectsProp) {
+  const navigate = useNavigate();
+
   return (
     <section className="relative w-full h-full overflow-hidden p-5">
       <motion.div
@@ -29,11 +35,32 @@ function Projects({ isMobile, isDarkMode }: ProjectsProp) {
         )}
       </motion.div>
 
-      <ProjectCard
-        isDarkMode={isDarkMode}
-        isMobile={isMobile}
-        projects={projects}
-      />
+      <div
+        className={`grid ${
+          isMobile ? "grid-cols-1" : "grid-cols-3"
+        } gap-5 text-white`}
+      >
+        {projects.slice(0, 3).map((project: Project, index: number) => (
+          <ProjectCard
+            isDarkMode={isDarkMode}
+            isMobile={isMobile}
+            project={project}
+            key={`project-${index}`}
+          />
+        ))}
+        <motion.div
+          variants={slideDown(1.2)}
+          className={`${
+            !isMobile && "col-span-3"
+          } flex justify-center items-center`}
+          onClick={() => {
+            navigate("/projects/all");
+            window.scrollTo({ top: 0 });
+          }}
+        >
+          <IconButton icon={<IconArrowRight />} text="Show More Projects" />
+        </motion.div>
+      </div>
     </section>
   );
 }